+++ source/main/map_editor/gui/GuiCom_Screen.cpp
@@ -99,7 +99,7 @@ void CGuiCom::InitGuiScreenRes()
 		std::vector<ScrRes> vRes;
 
 		const StringVector& videoModes = app->mRoot->getRenderSystem()->getConfigOptions()["Video Mode"].possibleValues;
-		for (int i=0; i < videoModes.size(); ++i)
+		for (size_t i=0; i < videoModes.size(); ++i)
 		{
 			String mode = videoModes[i];
 			StringUtil::trim(mode);
